requirements.txt
groq==0.5.0
langchain==0.1.16
langchain_groq==0.1.3
python-dotenv==1.0.1
streamlit==1.33.0
